+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [2.0.0] ** Breaking change **
+
+* Now `KeyboardActions` works on Dialogs
+* Add KeyboardActionsConfig to make parameters easily swappable
+* Add `FormKeyboardActions.setKeyboardActions` and `FormKeyboardActions` to allow changing the  config from anywhere in the child widget tree. (Check the sample)
+* Thanks @jayjwarrick for the contribution
+
 ## [1.0.4]
 
 * Added `enabled` attribute for KeyboardAction to skip the prev/next when the TextField is disabled